Move evaluate guess to turn file and test
